% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_rule.R
\name{wad_compute_target_storage}
\alias{wad_compute_target_storage}
\title{Compute target storage from elevation rule curve}
\usage{
wad_compute_target_storage(
  elev_df = NULL,
  min_storage_acft,
  max_storage_acft,
  elev_file = NULL
)
}
\arguments{
\item{elev_df}{Data frame with Target_Elev (ft), optional}

\item{min_storage_acft}{Minimum conservation storage (ac-ft)}

\item{max_storage_acft}{Maximum conservation storage (ac-ft)}

\item{elev_file}{Optional path to user elevation CSV (overrides default)}
}
\value{
Data frame with target storage (MCM)
}
\description{
Compute target storage from elevation rule curve
}
