% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_policy.R
\name{wad_release_policy}
\alias{wad_release_policy}
\title{WaDorule release policy with optional release plot}
\usage{
wad_release_policy(
  df,
  seasonal_weights,
  Smin,
  Smax,
  Qmin,
  Qmax,
  rule = "linear",
  plot = FALSE
)
}
\arguments{
\item{df}{Data frame with Date and Storage (MCM)}

\item{seasonal_weights}{Named vector of monthly inflow weights}

\item{Smin}{Minimum storage (MCM)}

\item{Smax}{Maximum storage (MCM)}

\item{Qmin}{Minimum release (m3/s)}

\item{Qmax}{Maximum release (m3/s)}

\item{rule}{linear or logistic}

\item{plot}{Logical; if TRUE, generates a daily release plot}
}
\value{
Data frame with Date, Storage, Month, Release
#' #' @examples
elev_file <- system.file("extdata", "Falls_Lake_Dam_Elevation.csv", package = "WaDorule")
storage_df <- wad_compute_target_storage(
  elev_df = read.csv(elev_file),
  min_storage_acft = 50000,
  max_storage_acft = 200000
)
}
\description{
Computes daily releases from a storage rule curve and monthly seasonal weights.
}
