% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_weights.R
\name{wad_seasonal_weights}
\alias{wad_seasonal_weights}
\title{Compute seasonal inflow weights}
\usage{
wad_seasonal_weights(inflow_df, inflow_col = "Q_upstream")
}
\arguments{
\item{inflow_df}{Data frame with Date and inflow}

\item{inflow_col}{Column name of inflow}
}
\value{
Named vector of monthly weights
#' @examples
elev_file <- system.file("extdata", "Falls_Lake_Dam_Elevation.csv", package = "WaDorule")
storage_df <- wad_compute_target_storage(
  elev_df = read.csv(elev_file),
  min_storage_acft = 50000,
  max_storage_acft = 200000
)
}
\description{
Compute seasonal inflow weights
}
